Support for YAML configuration files
JSON is still supported too.

- Moved modile compilation code to urbansim/modelcompile.py
- Broke model config loading code out into a separate function
    - Detects whether file is JSON or YAML, uses appropriate loader
- Modified gen_model to only take dicts of configuration parameters
- compilecli.py code now takes care of loading the configs from the files
