+++ urbansim/urbansim/tests/test_modelcompile.py
@@ -1,16 +1,11 @@
-"""
-Tests of the command line interface for generating model Python files.
-
-"""
 import os.path
 import shutil
 import tempfile
 
-import pytest
-import yaml
 import simplejson as json
+import yaml
 
-from .. import compilecli
+from .. import modelcompile
 
 TEST_CONFIG = {
     'growth_rate': 0.05,
@@ -30,34 +25,25 @@ def teardown_module(module):
     shutil.rmtree(module.TEST_DIR)
 
 
-@pytest.fixture(autouse=True)
-def temp_data_dir(monkeypatch):
-    monkeypatch.setenv('DATA_HOME', TEST_DIR)
-
-
-def test_model_save_with_dict():
-    compilecli.model_save(TEST_CONFIG)
-    assert os.path.exists(os.path.join(TEST_DIR, 'models', 'autorun_run.py'))
-
-
-def test_model_save_with_file_json():
+def test_load_config_json():
     test_file = os.path.join(TEST_DIR, 'test_config.json')
 
     with open(test_file, 'w') as f:
         json.dump(TEST_CONFIG, f)
 
-    compilecli.model_save(test_file)
-    assert os.path.exists(
-        os.path.join(TEST_DIR, 'models', 'test_config_run.py'))
+    config, basename = modelcompile.load_config(test_file)
+    assert config == TEST_CONFIG
+    assert basename == 'test_config.json'
 
 
-def test_model_save_with_file_yaml():
+def test_load_config_yaml():
     test_file = os.path.join(TEST_DIR, 'test_config.yaml')
 
     with open(test_file, 'w') as f:
         yaml.dump(
             TEST_CONFIG, f, default_flow_style=False, indent=4, width=50)
 
-    compilecli.model_save(test_file)
-    assert os.path.exists(
-        os.path.join(TEST_DIR, 'models', 'test_config_run.py'))
+    config, basename = modelcompile.load_config(test_file)
+    assert config == TEST_CONFIG
+    assert basename == 'test_config.yaml'
+
